refactor(ota): Refactor / cleanup OTA functions
