+++ code/components/jomjol_fileserver_ota/server_ota.cpp
@@ -22,7 +22,7 @@ https://docs.espressif.com/projects/esp-idf/en/latest/esp32/migration-guides/rel
 #include "esp_app_format.h"
 
 #include "MainFlowControl.h"
-#include "server_file.h"
+#include "server_help.h"
 #include "server_GPIO.h"
 
 #ifdef ENABLE_MQTT
@@ -38,12 +38,9 @@ https://docs.espressif.com/projects/esp-idf/en/latest/esp32/migration-guides/rel
 
 static const char *TAG = "SERVER_OTA";
 
-/*an ota data write buffer ready to write to the flash*/
+std::string file_name_update;
+// An ota data write buffer ready to write to the flash
 static char ota_write_data[SERVER_OTA_SCRATCH_BUFSIZE + 1] = { 0 };
-esp_err_t handler_reboot(httpd_req_t *req);
-static bool ota_update_task(std::string fn);
-std::string _file_name_update;
-bool initial_setup = false;
 
 
 static void infinite_loop(void)
@@ -57,80 +54,7 @@ static void infinite_loop(void)
 }
 
 
-void task_do_Update_ZIP(void *pvParameter)
-{
-    StatusLED(AP_OR_OTA, 1, true);  // Signaling an OTA update
-    
-    std::string filetype = toUpper(getFileType(_file_name_update));
-
-  	LogFile.WriteToFile(ESP_LOG_INFO, TAG, "File: " + _file_name_update + " Filetype: " + filetype);
-
-
-    if (filetype == "ZIP")
-    {
-        std::string in, out, outbin, zw, retfirmware;
-
-        out = "/sdcard/html";
-        outbin = "/sdcard/firmware";
-
-        retfirmware = unzip_new(_file_name_update, out+"/", outbin+"/", "/sdcard/", initial_setup);
-    	LogFile.WriteToFile(ESP_LOG_INFO, TAG, "Files unzipped.");
-
-        if (retfirmware.length() > 0)
-        {
-        	LogFile.WriteToFile(ESP_LOG_INFO, TAG, "Found firmware.bin");
-            ota_update_task(retfirmware);
-        }
-
-        LogFile.WriteToFile(ESP_LOG_INFO, TAG, "Trigger reboot due to firmware update");
-        doRebootOTA();
-    } else if (filetype == "BIN")
-    {
-       	LogFile.WriteToFile(ESP_LOG_INFO, TAG, "Do firmware update - file: " + _file_name_update);
-        ota_update_task(_file_name_update);
-        LogFile.WriteToFile(ESP_LOG_INFO, TAG, "Trigger reboot due to firmware update");
-        doRebootOTA();
-    }
-    else
-    {
-    	LogFile.WriteToFile(ESP_LOG_ERROR, TAG, "Only ZIP-Files support for update during startup");
-    }
-}
-
-
-void CheckUpdate()
-{
- 	FILE *pfile;
-    if ((pfile = fopen("/sdcard/update.txt", "r")) == NULL)
-    {
-		LogFile.WriteToFile(ESP_LOG_DEBUG, TAG, "No pending update");
-        return;
-	}
-
-	char zw[256] = "";
-	fgets(zw, sizeof(zw), pfile);
-    _file_name_update = std::string(zw);
-    if (fgets(zw, sizeof(zw), pfile))
-	{
-		std::string _szw = std::string(zw);
-        if (_szw == "init")
-        {
-       		LogFile.WriteToFile(ESP_LOG_INFO, TAG, "Inital Setup triggered");
-        }
-	}
-
-    fclose(pfile);
-    DeleteFile("/sdcard/update.txt");   // Prevent Boot Loop!!!
-	LogFile.WriteToFile(ESP_LOG_INFO, TAG, "Start update process (" + _file_name_update + ")");
-
-
-    xTaskCreate(&task_do_Update_ZIP, "task_do_Update_ZIP", configMINIMAL_STACK_SIZE * 35, NULL, tskIDLE_PRIORITY+1, NULL);
-    while(1) { // wait until reboot within task_do_Update_ZIP
-        vTaskDelay(1000 / portTICK_PERIOD_MS);
-    }
-}
-
-
+// OTA update: 3rd step
 static bool ota_update_task(std::string fn)
 {
     esp_err_t err;
@@ -151,15 +75,14 @@ static bool ota_update_task(std::string fn)
     ESP_LOGI(TAG, "Running partition type %d subtype %d (offset 0x%08x)",
              running->type, running->subtype, (unsigned int)running->address);
 
-
     update_partition = esp_ota_get_next_update_partition(NULL);
     ESP_LOGI(TAG, "Writing to partition subtype %d at offset 0x%x",
              update_partition->subtype, (unsigned int)update_partition->address);
-//    assert(update_partition != NULL);
+    //assert(update_partition != NULL);
 
     int binary_file_length = 0;
 
-    // deal with all receive packet 
+    //deal with all receive packet 
     bool image_header_was_checked = false;
 
     int data_read;     
@@ -176,7 +99,8 @@ static bool ota_update_task(std::string fn)
         if (data_read < 0) {
             LogFile.WriteToFile(ESP_LOG_ERROR, TAG, "Error: SSL data read error");
             return false;
-        } else if (data_read > 0) {
+        }
+        else if (data_read > 0) {
             if (image_header_was_checked == false) {
                 esp_app_desc_t new_app_info;
                 if (data_read > sizeof(esp_image_header_t) + sizeof(esp_image_segment_header_t) + sizeof(esp_app_desc_t)) {
@@ -206,12 +130,12 @@ static bool ota_update_task(std::string fn)
                         }
                     }
 
-/*
+                    /*
                     if (memcmp(new_app_info.version, running_app_info.version, sizeof(new_app_info.version)) == 0) {
                         LogFile.WriteToFile(ESP_LOG_WARN, TAG, "Current running version is the same as a new. We will not continue the update");
                         infinite_loop();
                     }
-*/
+                    */
                     image_header_was_checked = true;
 
                     err = esp_ota_begin(update_partition, OTA_SIZE_UNKNOWN, &update_handle);
@@ -220,18 +144,20 @@ static bool ota_update_task(std::string fn)
                         return false;
                     }
                     ESP_LOGI(TAG, "esp_ota_begin succeeded");
-                } else {
+                }
+                else {
                     LogFile.WriteToFile(ESP_LOG_ERROR, TAG, "received package is not fit len");
                     return false;
                 }
             }            
-            err = esp_ota_write( update_handle, (const void *)ota_write_data, data_read);
+            err = esp_ota_write(update_handle, (const void *)ota_write_data, data_read);
             if (err != ESP_OK) {
                 return false;
             }
             binary_file_length += data_read;
             ESP_LOGD(TAG, "Written image length %d", binary_file_length);
-        } else if (data_read == 0) {
+        }
+        else if (data_read == 0) {
            //
            // * As esp_http_client_read never returns negative error code, we rely on
            // * `errno` to check for underlying transport connectivity closure if any
@@ -259,16 +185,83 @@ static bool ota_update_task(std::string fn)
     err = esp_ota_set_boot_partition(update_partition);
     if (err != ESP_OK) {
         LogFile.WriteToFile(ESP_LOG_ERROR, TAG, "esp_ota_set_boot_partition failed (" + std::string(esp_err_to_name(err)) + ")");
-
     }
-//    ESP_LOGI(TAG, "Prepare to restart system");
-//    esp_restart();
 
     return true ;
 }
 
 
-static void print_sha256 (const uint8_t *image_hash, const char *label)
+// OTA update: 2nd step
+void task_do_update(void *pvParameter)
+{
+  	StatusLED(AP_OR_OTA, 1, true);  // Signaling an OTA update  
+
+    std::string filetype = toUpper(getFileType(file_name_update));
+    LogFile.WriteToFile(ESP_LOG_INFO, TAG, "File name: " + file_name_update + " | File type: " + filetype);
+
+    if (filetype == "ZIP") {
+        LogFile.WriteToFile(ESP_LOG_INFO, TAG, "Processing ZIP file: " + file_name_update);
+        std::string out, outbin, retfirmware;
+
+        out = "/sdcard/html";
+        outbin = "/sdcard/firmware";
+
+        retfirmware = unzip_new(file_name_update, out + "/", outbin + "/", "/sdcard/", false);
+    	LogFile.WriteToFile(ESP_LOG_INFO, TAG, "Files unzipped");
+
+        if (retfirmware.length() > 0) {
+        	LogFile.WriteToFile(ESP_LOG_INFO, TAG, "Found firmware.bin");
+            ota_update_task(retfirmware);
+        }
+
+        LogFile.WriteToFile(ESP_LOG_INFO, TAG, "Trigger reboot due to firmware update");
+        doRebootOTA();
+    }
+    else if (filetype == "BIN") {
+       	LogFile.WriteToFile(ESP_LOG_INFO, TAG, "Processing BIN file: " + file_name_update);
+        ota_update_task(file_name_update);
+        LogFile.WriteToFile(ESP_LOG_INFO, TAG, "Trigger reboot due to firmware update");
+        doRebootOTA();
+    }
+    else {
+    	LogFile.WriteToFile(ESP_LOG_ERROR, TAG, "Only ZIP or BIN files are supported for firmware update during startup");
+    }
+}
+
+
+// OTA update: 1st step
+void CheckUpdate()
+{
+ 	FILE *pfile;
+    if ((pfile = fopen("/sdcard/update.txt", "r")) == NULL) {
+		LogFile.WriteToFile(ESP_LOG_INFO, TAG, "No pending update");
+        return;
+	}
+
+	char zw[256] = "";
+	fgets(zw, sizeof(zw), pfile);
+    file_name_update = std::string(zw);
+    if (fgets(zw, sizeof(zw), pfile)) {
+        if (std::string(zw) == "init") {
+       		LogFile.WriteToFile(ESP_LOG_INFO, TAG, "Inital Setup triggered");
+        }
+	}
+
+    fclose(pfile);
+    DeleteFile("/sdcard/update.txt");   // Prevent Boot Loop!!!
+
+	LogFile.WriteToFile(ESP_LOG_INFO, TAG, "Prepare update process | File: " + file_name_update);
+    xTaskCreate(&task_do_update, "task_do_update", configMINIMAL_STACK_SIZE * 35, NULL, tskIDLE_PRIORITY+1, NULL);
+
+    while(1) { // wait until reboot within task_do_update
+        vTaskDelay(1000 / portTICK_PERIOD_MS);
+    }
+}
+
+
+//****************************************************************************
+#ifdef CONFIG_BOOTLOADER_APP_ROLLBACK_ENABLE
+static void print_sha256(const uint8_t *image_hash, const char *label)
 {
     char hash_print[HASH_LEN * 2 + 1];
     hash_print[HASH_LEN * 2] = 0;
@@ -285,9 +278,10 @@ static bool diagnostic(void)
 }
 
 
-void CheckOTAUpdate(void)
+// OTA status check is only needed if sdkconfig flag CONFIG_BOOTLOADER_APP_ROLLBACK_ENABLE is set
+void CheckOTAStatus(void)
 {
-    ESP_LOGI(TAG, "Start CheckOTAUpdateCheck");
+    ESP_LOGI(TAG, "Check OTAStatus");
 
     uint8_t sha_256[HASH_LEN] = { 0 };
     esp_partition_t partition;
@@ -312,40 +306,10 @@ void CheckOTAUpdate(void)
 
     const esp_partition_t *running = esp_ota_get_running_partition();
     esp_ota_img_states_t ota_state;
-    esp_err_t res_stat_partition = esp_ota_get_state_partition(running, &ota_state);
-    switch (res_stat_partition)
-    {
-        case ESP_OK:
-            ESP_LOGD(TAG, "CheckOTAUpdate Partition: ESP_OK");
-            if (esp_ota_get_state_partition(running, &ota_state) == ESP_OK) {
-                if (ota_state == ESP_OTA_IMG_PENDING_VERIFY) {
-                    // run diagnostic function
-                    bool diagnostic_is_ok = diagnostic();
-                    if (diagnostic_is_ok) {
-                        ESP_LOGI(TAG, "Diagnostics completed successfully! Continuing execution");
-                        esp_ota_mark_app_valid_cancel_rollback();
-                    } else {
-                        LogFile.WriteToFile(ESP_LOG_ERROR, TAG, "Diagnostics failed! Start rollback to the previous version");
-                        esp_ota_mark_app_invalid_rollback_and_reboot();
-                    }
-                }
-            }            
-            break;
-        case ESP_ERR_INVALID_ARG:
-            ESP_LOGD(TAG, "CheckOTAUpdate Partition: ESP_ERR_INVALID_ARG");
-            break;
-        case ESP_ERR_NOT_SUPPORTED:
-            ESP_LOGD(TAG, "CheckOTAUpdate Partition: ESP_ERR_NOT_SUPPORTED");
-            break;
-        case ESP_ERR_NOT_FOUND:
-            ESP_LOGD(TAG, "CheckOTAUpdate Partition: ESP_ERR_NOT_FOUND");
-            break;
-    }
     if (esp_ota_get_state_partition(running, &ota_state) == ESP_OK) {
         if (ota_state == ESP_OTA_IMG_PENDING_VERIFY) {
             // run diagnostic function
-            bool diagnostic_is_ok = diagnostic();
-            if (diagnostic_is_ok) {
+            if (diagnostic()) {
                 ESP_LOGI(TAG, "Diagnostics completed successfully! Continuing execution");
                 esp_ota_mark_app_valid_cancel_rollback();
             } else {
@@ -355,142 +319,79 @@ void CheckOTAUpdate(void)
         }
     }
 }
+#endif
+//****************************************************************************
 
 
 esp_err_t handler_ota_update(httpd_req_t *req)
 {
-#ifdef DEBUG_DETAIL_ON     
-    LogFile.WriteHeapInfo("handler_ota_update - Start");    
-#endif
-
     LogFile.WriteToFile(ESP_LOG_DEBUG, TAG, "handler_ota_update");
     char _query[200];
     char _filename[100];
     char _valuechar[30];    
     std::string fn = "/sdcard/firmware/";
-    bool _file_del = false;
     std::string _task = "";
+    bool _file_del = false;
 
-    if (httpd_req_get_url_query_str(req, _query, 200) == ESP_OK)
-    {
-        ESP_LOGD(TAG, "Query: %s", _query);
-        
-        if (httpd_query_key_value(_query, "task", _valuechar, 30) == ESP_OK)
-        {
-            ESP_LOGD(TAG, "task is found: %s", _valuechar);
+    httpd_resp_set_hdr(req, "Access-Control-Allow-Origin", "*");
+
+    if (httpd_req_get_url_query_str(req, _query, 200) == ESP_OK) {
+        //ESP_LOGD(TAG, "Query: %s", _query);    
+        if (httpd_query_key_value(_query, "task", _valuechar, sizeof(_valuechar)) == ESP_OK) {
+            //ESP_LOGD(TAG, "task is found: %s", _valuechar);
             _task = std::string(_valuechar);
         }
 
-        if (httpd_query_key_value(_query, "file", _filename, 100) == ESP_OK)
-        {
+        if (httpd_query_key_value(_query, "file", _filename, sizeof(_filename)) == ESP_OK) {
             fn.append(_filename);
-            ESP_LOGD(TAG, "File: %s", fn.c_str());
+            //ESP_LOGD(TAG, "File: %s", fn.c_str());
         }
-        if (httpd_query_key_value(_query, "delete", _filename, 100) == ESP_OK)
-        {
+
+        if (httpd_query_key_value(_query, "delete", _filename, sizeof(_filename)) == ESP_OK) {
             fn.append(_filename);
             _file_del = true;
-            ESP_LOGD(TAG, "Delete Default File: %s", fn.c_str());
+            //ESP_LOGD(TAG, "Delete Default File: %s", fn.c_str());
         }
-
     }
 
-    if (_task.compare("emptyfirmwaredir") == 0)
-    {
-        ESP_LOGD(TAG, "Start empty directory /firmware");
+    if (_task.compare("emptyfirmwaredir") == 0) {
         delete_all_in_directory("/sdcard/firmware");
-        std::string zw = "firmware directory deleted - v2\n";
-        ESP_LOGD(TAG, "%s", zw.c_str());
-        printf("Ausgabe: %s\n", zw.c_str());
-    
-        httpd_resp_set_hdr(req, "Access-Control-Allow-Origin", "*");
-        httpd_resp_send(req, zw.c_str(), strlen(zw.c_str())); 
-        /* Respond with an empty chunk to signal HTTP response completion */
-        httpd_resp_send_chunk(req, NULL, 0);  
-
-        ESP_LOGD(TAG, "Done empty directory /firmware");
+        httpd_resp_sendstr(req, "Directory /sdcard/firmware deleted"); 
         return ESP_OK;
     }
 
-    if (_task.compare("update") == 0)
-    {
+    else if (_task.compare("update") == 0) {
         std::string filetype = toUpper(getFileType(fn));
-        if (filetype.length() == 0)
-        {
-            std::string zw = "Update failed - no file specified (zip, bin, tfl, tlite)";
-            httpd_resp_sendstr_chunk(req, zw.c_str());
-            httpd_resp_sendstr_chunk(req, NULL);  
-            return ESP_OK;        
-        }
-
-
-        if ((filetype == "TFLITE") || (filetype == "TFL"))
-        {
+        if ((filetype == "TFLITE") || (filetype == "TFL")) {
             std::string out = "/sdcard/config/" + getFileFullFileName(fn);
             DeleteFile(out);
             CopyFile(fn, out);
             DeleteFile(fn);
 
-            const char*  resp_str = "Neural Network File copied.";
-            httpd_resp_sendstr_chunk(req, resp_str);
-            httpd_resp_sendstr_chunk(req, NULL);  
+            LogFile.WriteToFile(ESP_LOG_INFO, TAG, "TFLITE/TFL file: Update completed");
+            httpd_resp_sendstr(req, "Neural network file updated. No reboot required");
             return ESP_OK;
         }
 
+        else if ((filetype == "ZIP") || (filetype == "BIN")) {
+            LogFile.WriteToFile(ESP_LOG_INFO, TAG, "ZIP/BIN file: Reboot required to update");
 
-        if ((filetype == "ZIP") || (filetype == "BIN"))
-        {
-           	FILE *pfile;
-            LogFile.WriteToFile(ESP_LOG_INFO, TAG, "Update for reboot");
-            pfile = fopen("/sdcard/update.txt", "w");
+            FILE *pfile = fopen("/sdcard/update.txt", "w");
             fwrite(fn.c_str(), fn.length(), 1, pfile);
             fclose(pfile);
 
-            std::string zw = "reboot\n";
-            httpd_resp_sendstr_chunk(req, zw.c_str());
-            httpd_resp_sendstr_chunk(req, NULL);  
-            ESP_LOGD(TAG, "Send reboot");
-            return ESP_OK;                
-
+            httpd_resp_sendstr(req, "reboot"); // String needs to be started with "reboot" -> Trigger reboot from WebUI
+            return ESP_OK;  
         }
 
-/*
-        if (filetype == "BIN")
-        {
-            const char* resp_str; 
-
-            DeleteMainFlowTask();
-            gpio_handler_deinit();
-            if (ota_update_task(fn))
-            {
-                std::string zw = "reboot\n";
-                httpd_resp_sendstr_chunk(req, zw.c_str());
-                httpd_resp_sendstr_chunk(req, NULL);  
-                ESP_LOGD(TAG, "Send reboot");
-                return ESP_OK;                
-            }
-
-            resp_str = "Error during Firmware Update!!!\nPlease check output of console.";
-            httpd_resp_send(req, resp_str, strlen(resp_str));  
-
-            #ifdef DEBUG_DETAIL_ON 
-                LogFile.WriteHeapInfo("handler_ota_update - Done");    
-            #endif
-
-            return ESP_OK;
-        }
-*/
-
-        std::string zw = "Update failed - no valid file specified (zip, bin, tfl, tlite)";
-        httpd_resp_sendstr_chunk(req, zw.c_str());
-        httpd_resp_sendstr_chunk(req, NULL);  
-        return ESP_OK;        
+        std::string zw = "task=update: No valid file (.zip, .bin, .tfl, .tlite)";
+        LogFile.WriteToFile(ESP_LOG_ERROR, TAG, zw);
+        httpd_resp_send_err(req, HTTPD_400_BAD_REQUEST, zw.c_str()); 
+        return ESP_FAIL;        
     }
 
-
-    if (_task.compare("unziphtml") == 0)
-    {
-        ESP_LOGD(TAG, "Task unziphtml");
+    else if (_task.compare("unziphtml") == 0) {
+        //ESP_LOGD(TAG, "Task unziphtml");
         std::string in, out, zw;
 
         in = "/sdcard/firmware/html.zip";
@@ -498,79 +399,44 @@ esp_err_t handler_ota_update(httpd_req_t *req)
 
         delete_all_in_directory(out);
 
-        unzip(in, out+"/");
-        zw = "Web Interface Update Successfull!\nNo reboot necessary";
-        httpd_resp_send(req, zw.c_str(), strlen(zw.c_str()));
-        httpd_resp_sendstr_chunk(req, NULL);  
+        unzip(in, out + "/");
+
+        LogFile.WriteToFile(ESP_LOG_INFO, TAG, "Web interface: Update completed");
+        httpd_resp_sendstr(req, "Web interface: Update completed. No reboot required");
         return ESP_OK;        
     }
 
-    if (_file_del)
-    {
-        ESP_LOGD(TAG, "Delete !! _file_del: %s", fn.c_str());
-        struct stat file_stat;
-        int _result = stat(fn.c_str(), &file_stat);
-        ESP_LOGD(TAG, "Ergebnis %d\n", _result);
-        if (_result == 0) {
-            ESP_LOGD(TAG, "Deleting file: %s", fn.c_str());
-            /* Delete file */
-            unlink(fn.c_str());
+    if (_file_del) {
+        if(!DeleteFile(fn)) {
+            LogFile.WriteToFile(ESP_LOG_ERROR, TAG, "Deletetion failed. File does not exist: " + fn);
+            httpd_resp_send_err(req, HTTPD_404_NOT_FOUND, "File deletion failed"); 
+            return ESP_FAIL;
         }
-        else
-        {
-            LogFile.WriteToFile(ESP_LOG_ERROR, TAG, "File does not exist: " + fn);
-        }
-        /* Respond with an empty chunk to signal HTTP response completion */
-        std::string zw = "file deleted\n";
-        ESP_LOGD(TAG, "%s", zw.c_str());
-        httpd_resp_send(req, zw.c_str(), strlen(zw.c_str()));
-        httpd_resp_send_chunk(req, NULL, 0);
+        
+        std::string zw = "File deleted: " + fn;
+        LogFile.WriteToFile(ESP_LOG_INFO, TAG, zw);
+        httpd_resp_sendstr(req, zw.c_str());
         return ESP_OK;
     }
 
-    std::string zw = "ota without parameter - should not be the case";
-    httpd_resp_set_hdr(req, "Access-Control-Allow-Origin", "*");
-    httpd_resp_send(req, zw.c_str(), strlen(zw.c_str())); 
-    httpd_resp_send_chunk(req, NULL, 0);  
-
-    LogFile.WriteToFile(ESP_LOG_ERROR, TAG, "ota without parameter - should not be the case");
-
-/*  
-    const char* resp_str;    
-
-    DeleteMainFlowTask();
-    gpio_handler_deinit();
-    if (ota_update_task(fn))
-    {
-        resp_str = "Firmware Update Successfull! You can restart now.";
-    }
-    else
-    {
-        resp_str = "Error during Firmware Update!!! Please check console output.";
-    }
-
-    httpd_resp_send(req, resp_str, strlen(resp_str));  
-*/
-
-    #ifdef DEBUG_DETAIL_ON 
-        LogFile.WriteHeapInfo("handler_ota_update - Done");    
-    #endif
-
-    return ESP_OK;
+    std::string zw = "No valid task/action: OTA handler called without any parameter";
+    LogFile.WriteToFile(ESP_LOG_ERROR, TAG, zw);
+    httpd_resp_send_err(req, HTTPD_400_BAD_REQUEST, zw.c_str()); 
+    return ESP_FAIL;
 }
 
 
 void hard_restart() 
 {
-  esp_task_wdt_config_t twdt_config = {
-    .timeout_ms = 1,
-    .idle_core_mask = (1 << portNUM_PROCESSORS) - 1,    // Bitmask of all cores
-    .trigger_panic = true,
-  };
-  ESP_ERROR_CHECK(esp_task_wdt_init(&twdt_config));
+    esp_task_wdt_config_t twdt_config = {
+        .timeout_ms = 1,
+        .idle_core_mask = (1 << portNUM_PROCESSORS) - 1,    // Bitmask of all cores
+        .trigger_panic = true,
+    };
+    ESP_ERROR_CHECK(esp_task_wdt_init(&twdt_config));
 
-  esp_task_wdt_add(NULL);
-  while(true);
+    esp_task_wdt_add(NULL);
+    while(true);
 }
 
 
@@ -645,10 +511,6 @@ void doRebootOTA()
 
 esp_err_t handler_reboot(httpd_req_t *req)
 {
-    #ifdef DEBUG_DETAIL_ON     
-        LogFile.WriteHeapInfo("handler_reboot - Start");
-    #endif    
-
     LogFile.WriteToFile(ESP_LOG_DEBUG, TAG, "handler_reboot");
     LogFile.WriteToFile(ESP_LOG_INFO, TAG, "System will restart within 5 sec");
 
@@ -666,10 +528,6 @@ esp_err_t handler_reboot(httpd_req_t *req)
     httpd_resp_send(req, response.c_str(), strlen(response.c_str()));
     
     doReboot();
-
-    #ifdef DEBUG_DETAIL_ON 
-        LogFile.WriteHeapInfo("handler_reboot - Done");    
-    #endif
 
     return ESP_OK;
 }
