+++ code/components/jomjol_fileserver_ota/server_ota.h
@@ -4,11 +4,15 @@
 #include <esp_http_server.h>
 
 
-void register_server_ota_sdcard_uri(httpd_handle_t server);
+void CheckUpdate();
+#ifdef CONFIG_BOOTLOADER_APP_ROLLBACK_ENABLE
 void CheckOTAUpdate();
+#endif
+
 void doReboot();
 void doRebootOTA();
 void hard_restart();
-void CheckUpdate();
+
+void register_server_ota_sdcard_uri(httpd_handle_t server);
 
 #endif //SERVEROTA_H
