+++ code/test/test_suite_flowcontroll.cpp
@@ -158,8 +158,11 @@ extern "C" void app_main()
     // Check for updates before start testing
     // It is possbile to update thr firmware also by placing 'firmware.bin' to '/sdcard/firmware' and
     // file 'update.txt' with content '/sdcard/firmware/firmware.bin' to sd card root folder.
+    // Note: OTA Status check only necessary if OTA rollback feature is enabled
     // ********************************************
-    CheckOTAUpdate();
+    #ifdef CONFIG_BOOTLOADER_APP_ROLLBACK_ENABLE
+    CheckOTAStatus();
+    #endif
     CheckUpdate();
